# `Development`: Remove sub-queries for filtering and ordering posts