--- a/src/main/java/de/tum/in/www1/artemis/service/metis/ReactionService.java
+++ b/src/main/java/de/tum/in/www1/artemis/service/metis/ReactionService.java
@@ -18,6 +18,8 @@ public class ReactionService {
 
     private static final String METIS_REACTION_ENTITY_NAME = "posting reaction";
 
+    private static final String VOTE_EMOJI_ID = "heavy_plus_sign";
+
     private final UserRepository userRepository;
 
     private final CourseRepository courseRepository;
@@ -68,6 +70,12 @@ public class ReactionService {
             reaction.setPost(post);
             // save reaction
             savedReaction = reactionRepository.save(reaction);
+
+            if (reaction.getEmojiId().equals(VOTE_EMOJI_ID)) {
+                // increase voteCount of post needed for sorting
+                post.setVoteCount(post.getVoteCount() != null ? post.getVoteCount() + 1 : 1);
+            }
+
             // save post
             postService.updateWithReaction(post, reaction, courseId);
         }
@@ -105,6 +113,14 @@ public class ReactionService {
             updatedPost = reaction.getPost();
             updatedPost.setConversation(mayInteractWithConversationIfConversationMessage(user, updatedPost));
             updatedPost.removeReaction(reaction);
+
+            if (reaction.getEmojiId().equals(VOTE_EMOJI_ID)) {
+                // decrease voteCount of post needed for sorting
+                updatedPost.setVoteCount(updatedPost.getVoteCount() - 1);
+
+                // save post
+                postService.updateWithReaction(updatedPost, reaction, courseId);
+            }
         }
         else {
             AnswerPost updatedAnswerPost = reaction.getAnswerPost();