--- a/src/main/java/de/tum/in/www1/artemis/service/metis/ReactionService.java
+++ b/src/main/java/de/tum/in/www1/artemis/service/metis/ReactionService.java
@@ -113,7 +113,7 @@ public class ReactionService {
             updatedPost = reaction.getPost();
             updatedPost.setConversation(mayInteractWithConversationIfConversationMessage(user, updatedPost));
 
-            if (reaction.getEmojiId().equals(VOTE_EMOJI_ID)) {
+            if (VOTE_EMOJI_ID.equals(reaction.getEmojiId())) {
                 // decrease voteCount of post needed for sorting
                 updatedPost.setVoteCount(updatedPost.getVoteCount() - 1);
             }