--- a/src/main/java/de/tum/in/www1/artemis/service/metis/ReactionService.java
+++ b/src/main/java/de/tum/in/www1/artemis/service/metis/ReactionService.java
@@ -18,6 +18,7 @@ public class ReactionService {
 
     private static final String METIS_REACTION_ENTITY_NAME = "posting reaction";
 
+    // constant must be same as it is in the client (metis.util.ts#28)
     private static final String VOTE_EMOJI_ID = "heavy_plus_sign";
 
     private final UserRepository userRepository;