--- a/src/test/java/de/tum/in/www1/artemis/metis/AnswerPostIntegrationTest.java
+++ b/src/test/java/de/tum/in/www1/artemis/metis/AnswerPostIntegrationTest.java
@@ -330,9 +330,9 @@ class AnswerPostIntegrationTest extends AbstractSpringIntegrationBambooBitbucket
         database.assertSensitiveInformationHidden(returnedPosts);
 
         int numberOfMaxAnswersSeenOnAnyPost = 0;
-        for (int i = 0; i < returnedPosts.size(); i++) {
-            assertThat(returnedPosts.get(i).getAnswers().size()).isGreaterThanOrEqualTo(numberOfMaxAnswersSeenOnAnyPost);
-            numberOfMaxAnswersSeenOnAnyPost = returnedPosts.get(i).getAnswers().size();
+        for (Post post : returnedPosts) {
+            assertThat(post.getAnswers().size()).isGreaterThanOrEqualTo(numberOfMaxAnswersSeenOnAnyPost);
+            numberOfMaxAnswersSeenOnAnyPost = post.getAnswers().size();
         }
     }