<!-- Thanks for contributing to Artemis! Before you submit your pull request, please make sure to check the following boxes by putting an x in the [ ] (don't: [x ], [ x], do: [x]) -->
<!-- If your pull request is not ready for review yet, create a draft pull request! -->

### Checklist
#### General
<!-- You only need to choose one of the first two check items: Generally, test on the test servers. -->
<!-- If it's only a small change, testing it locally is acceptable and you may remove the first checkmark. If you are unsure, please test on the test servers. -->
- [x] I tested **all** changes and their related features with **all** corresponding user types on a test server.
- [x] Language: I followed the [guidelines for inclusive, diversity-sensitive, and appreciative language](https://docs.artemis.ase.in.tum.de/dev/guidelines/language-guidelines/).
- [x] I chose a title conforming to the [naming conventions for pull requests](https://artemis-platform.readthedocs.io/en/latest/dev/guidelines/development-process.html#naming-conventions-for-github-pull-requests).
#### Server
- [x] I followed the [coding and design guidelines](https://docs.artemis.ase.in.tum.de/dev/guidelines/server/).
- [x] I added multiple integration tests (Spring) related to the features (with a high test coverage).
- [x] I implemented the changes with a good performance and prevented too many database calls.
- [x] I documented the Java code using JavaDoc style.


### Motivation and Context
While fetching course posts, we previously generated sub-queries to filter and sort posts, in accordance with many variant properties. The use of sub-queries is a bad programming habit, as they drastically reduce performance during database calls.

Also, due to the incompatibility of the DISTINCT and ORDER BY keywords used in Post queries, we previously had to disable server tests related to ordering posts by creation date, vote count, and answer count.

### Description
To remove the sub-queries while preserving the rich filtering and sorting of posts, we introduce three new columns to the post table; resolved, answer_count, and vote_count respectively. By pre-populating these columns for all existing Posts and accounting them with respect to the operations Post go under, we are able to avoid sub-queries and instead utilize the information stored in these columns to realize the filtering and sorting operations of Posts.

To re-enable the previously disabled post-server tests, we remove the DISTINCT keyword and instead use GROUP BY (Post.ID) to achieve the same result.

### Performance Differences
Measurements were performed on a dataset with 237 Posts, 299 Reactions, and 299 AnswerPosts.
| Sorting/Filtering Strategy | Previous Implementation | New Implementation|
|------------|-------:|-----:|
| Unresolved | 774ms | 203ms |
| Sort by Vote Count ASC |  194ms | 168ms |
| Sort by Vote Count DESC | 244ms | 192ms |
| Sort by Answer Count ASC | 174ms | 170ms |
| Sort by Answer Count DESC | 1.07s | 995ms |

[PerformanceMeasurements.zip](https://github.com/ls1intum/Artemis/files/9998019/PerformanceMeasurements.zip)

### Steps for Testing
<!-- Please describe in detail how the reviewer can test your changes. -->
Prerequisites:
- 1 Tutor
- 1 Course with discussions enabled

1. Log in to Artemis
2. Navigate to Course Communication Page
3. Verify that filters work as before
4. Make sure that posts are ordered correctly with respect to date, votes, and number of answers.
5. Alter votes and number of answers of posts, make sure that they are re-ordered correctly after the changes.
6. Announcement posts are treated as resolved by default. Resolve a non-announcement post and make sure it is not displayed when filtering for unresolved posts.
7. Unresolve the previous post and make sure that it is displayed again while filtering for unresolved posts.
8. Make sure that no Post is displayed more than once. (check Post ID next to Post title)


### Review Progress
<!-- Each Pull Request should be reviewed by at least two other developers. The code, the functionality (= manual test) and the exam mode need to be reviewed. -->
<!-- The reviewer or author check the following boxes depending on what was reviewed or tested. All boxes should be checked before merge. -->
<!-- You can add additional checkboxes if it makes sense to only review parts of the code or functionality. -->
<!-- When changes are pushed, uncheck the affected boxes. (Not all changes require full re-reviews.) -->
<!-- All PRs that might affect the exam mode (e.g. change a client component that is also used in the exam mode) need an additional verification that the exam mode still works. -->

#### Code Review
- [x] Review 1
- [x] Review 2
#### Manual Tests
- [x] Test 1
- [x] Test 2


### Test Coverage
<!-- Please add the test coverages for all changed files here. You can see this when executing the tests locally (see build.gradle and package.json) or when looking into the corresponding Bamboo build plan. -->
<!-- Lines are the main reference but a significantly lower branch percentage can indicate missing edge cases in the tests. -->
| Class/File | Branch | Line |
|------------|-------:|-----:|
| PostService.java | 81% | 97% |
| AnswerPostService.Java | 81% | 79% |
| ReactionService.java | 100% | 100% |
| PostSpecs.java | 87% | 85% |
<!--
| ExerciseService.java | 85% | 77% |
| programming-exercise.component.ts | 13% | 95% |
-->